+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Improved
+- Use gum spinner for loading states (vm, run)
+- Better error messages when API not enabled or permission denied
+
 ## [1.2.1] - 2026-01-01
 
 ### Fixed